修复starter无法设置刷新周期问题

+++ dictionary/dictionary-core/src/main/java/com/github/softwarevax/dict/core/domain/DictionaryConfigure.java
@@ -21,7 +21,7 @@ public class DictionaryConfigure {
         return refreshInterval;
     }
 
-    public void setRefreshInterval(int refreshInterval) {
+    public void setRefreshInterval(long refreshInterval) {
         if(refreshInterval <= 5) {
             return;
         }
